Refactor tests folder
