+++ packages/e2e/tests/e2e/cards/dualBranding/dualBranding.clientScripts.js
@@ -1,7 +1,7 @@
 /**
  * Set koreanAuthenticationRequired & countryCode so KCP fields show at start
  */
-window.cardConfigObj = {
+window.cardConfig = {
     type: 'scheme',
     brands: ['mc', 'visa', 'amex', 'cartebancaire']
 };
