+++ packages/e2e/tests/e2e/cards/kcp/startWithKCP/startWithKCP.test.js
@@ -1,11 +1,11 @@
 import { Selector, ClientFunction } from 'testcafe';
 
-import { start } from '../commonUtils';
+import { start } from '../../../commonUtils';
 
-import { fillCardNumber, fillDateAndCVC } from './utils/cardUtils';
-import { fillTaxNumber, fillPwd } from './utils/kcpUtils';
+import { fillCardNumber, fillDateAndCVC } from '../../utils/cardUtils';
+import { fillTaxNumber, fillPwd } from '../../utils/kcpUtils';
 
-import { KOREAN_TEST_CARD, REGULAR_TEST_CARD, TEST_TAX_NUMBER_VALUE } from '../constants';
+import { KOREAN_TEST_CARD, REGULAR_TEST_CARD, TEST_TAX_NUMBER_VALUE } from '../../../constants';
 
 const passwordHolder = Selector('.card-field [data-cse="encryptedPassword"]');
 
@@ -19,15 +19,7 @@ const getCardState = ClientFunction((what, prop) => {
 
 const TEST_SPEED = 1;
 
-/**
- * NOTE: For tests to work with a config file the playground file needs to look for the window.cardConfigObj that the config file creates
- * e.g.
- *  window.card = checkout
- *      .create('card', window.cardConfigObj || {
- *          type: 'scheme',
- *          brands ...
- */
-fixture`Starting with KCP fields`.page`http://localhost:3020/cards/`.clientScripts('config/startWithKCP.js');
+fixture`Starting with KCP fields`.page`http://localhost:3020/cards/`.clientScripts('startWithKCP.clientScripts.js');
 
 // Green 1
 test(
