+++ packages/e2e/tests/e2e/cards/kcp/startWithoutKCP/startWithoutKCP.clientScripts.js
@@ -1,11 +1,10 @@
 /**
- * Set koreanAuthenticationRequired & countryCode so KCP fields show at start
+ * Set koreanAuthenticationRequired so KCP fields will show if a korean card number is entered
  */
-window.cardConfigObj = {
+window.cardConfig = {
     type: 'scheme',
     brands: ['mc', 'visa', 'amex', 'korean_local_card'],
     configuration: {
         koreanAuthenticationRequired: true
-    },
-    countryCode: 'kr'
+    }
 };
